+++ hutool-extra/src/main/java/cn/hutool/extra/tokenizer/TokenizerEngine.java
@@ -6,7 +6,7 @@ package cn.hutool.extra.tokenizer;
  * @author looly
  *
  */
-public interface Engine {
+public interface TokenizerEngine {
 	
 	/**
 	 * 文本分词处理接口，通过实现此接口完成分词，产生分词结果