+++ hutool-extra/src/main/java/cn/hutool/extra/tokenizer/TokenizerUtil.java
@@ -1,6 +1,6 @@
 package cn.hutool.extra.tokenizer;
 
-import cn.hutool.extra.tokenizer.engine.EngineFactory;
+import cn.hutool.extra.tokenizer.engine.TokenizerFactory;
 
 /**
  * 分词工具类
@@ -13,9 +13,9 @@ public class TokenizerUtil {
 	/**
 	 * 根据用户引入的分词引擎jar，自动创建对应的分词引擎对象
 	 * 
-	 * @return {@link Engine}
+	 * @return {@link TokenizerEngine}
 	 */
-	public static Engine createEngine() {
-		return EngineFactory.create();
+	public static TokenizerEngine createEngine() {
+		return TokenizerFactory.create();
 	}
 }