+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * 【core】        增加Simhash（用于海量文本去重）
 * 【extra】        增加分词封装，封装了ansj、HanLP、IKAnalyzer、Jcseg、Jieba、MMSeg的实现，统一了接口
 * 【core】        去除NumberUtil.parseInt和parseLong的8进制支持（issue#234@Github）
+* 【extra】        Template部分修改命名减少歧义（Engine->TemplateEngine，EngineFactory->TemplateFactory）
 
 ### Bug修复